requirements.txt
streamlit
pandas
numpy
request
datetime
